Successfully increased no. of queues.
Having issues with setting up mingw/msys to use make on windows.
Interesting error: Trying to call a function which took an int as it's
argument, without passing it the argument made the program crash.
Usually these sort of things are spotted by the compiler, but it didn't
spot it for me :(